+++ 3rdparty/MelodixAPI/client/MDGetRecommendResource_200_response_recommend_inner.h
@@ -40,8 +40,8 @@ public:
     void fromJsonObject(QJsonObject json) override;
     void fromJson(QString jsonString) override;
 
-    qint32 getId() const;
-    void setId(const qint32 &id);
+    double getId() const;
+    void setId(const double &id);
     bool is_id_Set() const;
     bool is_id_Valid() const;
 
@@ -101,7 +101,7 @@ public:
 private:
     void initializeModel();
 
-    qint32 m_id;
+    double m_id;
     bool m_id_isSet;
     bool m_id_isValid;
